freeing memory on successful exit, changed some filenames and their content
