+++ srcs/free_memory.c
@@ -1,34 +1,44 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   input_error2.c                                     :+:      :+:    :+:   */
+/*   free_memory.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: plang <plang@student.hive.fi>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/16 11:29:01 by plang             #+#    #+#             */
-/*   Updated: 2024/08/16 13:17:11 by plang            ###   ########.fr       */
+/*   Updated: 2024/08/16 16:08:12 by plang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/cub3d.h"
 
-void	ft_loadpng_fail(t_fcheck *check, t_map *map, int index)
+void	ft_strfree(char *str)
 {
-	while (index >= 0)
+	if (str)
 	{
-		mlx_delete_texture(map->wall.txt[index]);
-		index--;
+		free(str);
+		str = NULL;
 	}
-	error_inside_file(check, MLXLPFAIL);
+	else
+		return ;
 }
 
-void	ft_loadpng_fail2(t_map *map, int index, int imagef)
+void	ft_arrfree(char **arr)
+{
+	if (arr)
+	{
+		free(arr);
+		arr = NULL;
+	}
+	else
+		return ;
+}
+
+void	successful_exit(t_map *map, int index)
 {
 	int	i;
 
 	i = 0;
-	if (imagef == 1)
-		mlx_delete_texture(map->background_txt);
 	while (i < map->map_size)
 	{
 		ft_strfree(map->mapsave[i]);
@@ -40,10 +50,5 @@ void	ft_loadpng_fail2(t_map *map, int index, int imagef)
 		mlx_delete_texture(map->wall.txt[index]);
 		index--;
 	}
-	if (imagef == 0)
-		ft_putstr_fd(MLXLPFAIL, 2);
-	else if (imagef == 1)
-		ft_putstr_fd(MLXTXTFAIL, 2);
-	mlx_terminate(map->mlx);
-	exit (1);
+	mlx_delete_image(map->mlx, map->background);
 }
